Package all functions into a desktop app Organize project files.

+++ DataAccess/Model/EventLogDetail.cs
@@ -17,7 +17,7 @@ namespace Service.Model
 
         public string ProductName { get; set; }
 
-        public byte UnitStatusID { get; set; }
+        public int UnitStatusID { get; set; }
 
         public string UnitStatusDescription {
             get {
